+++ src/fields/fq.rs
@@ -11,7 +11,7 @@ use lazy_static::lazy_static;
 #[cfg(feature = "bits")]
 use ff::{FieldBits, PrimeFieldBits};
 
-use crate::arithmetic::{adc, mac, sbb, FieldExt, SqrtTableHelpers};
+use crate::arithmetic::{adc, mac, sbb, SqrtTableHelpers};
 
 #[cfg(feature = "sqrt-table")]
 use crate::arithmetic::SqrtTables;
@@ -726,8 +726,6 @@ impl SqrtTableHelpers for Fq {
         tmp.0[0] as u32
     }
 }
-
-impl FieldExt for Fq {}
 
 impl WithSmallOrderMulGroup<3> for Fq {
     const ZETA: Self = Fq::from_raw([